문단 추가하는 기능 구현 -컨트롤 부분과 메모 부분은 아직 구현하지 않음...

+++ src/kr/dogfoot/hwplib/object/docinfo/charshape/Ratios.java
@@ -161,7 +161,7 @@ public class Ratios {
 	 * 
 	 * @return 사용자 정의 글자 장평
 	 */
-	public int getUser() {
+	public short getUser() {
 		return array[6];
 	}
