+++ src/kr/dogfoot/hwplib/object/docinfo/charshape/RelativeSizes.java
@@ -161,7 +161,7 @@ public class RelativeSizes {
 	 * 
 	 * @return 사용자 정의 글자 장평
 	 */
-	public int getUser() {
+	public short getUser() {
 		return array[6];
 	}
